+++ streamlit_permalink/handlers/data_editor.py
@@ -1,4 +1,7 @@
-from typing import Any, Optional
+from ..constants import DATAEDITOR_COLUMN_CONFIG_PREFIX, DATAEDITOR_DATE_VALUE_PREFIX, DATAEDITOR_DATETIME_VALUE_PREFIX, DATAEDITOR_PREFIX, DATAEDITOR_TIME_VALUE_PREFIX
+from ..url_validators import validate_single_url_value
+from .handler import WidgetHandler
+from typing import Any
 from io import StringIO
 
 import streamlit as st
@@ -6,33 +9,31 @@ import pandas as pd
 
 from .handler import WidgetHandler
 
+# TODO: Update df to prefix STREAMLIT_PERMALINK_TIME in front of time values ratehr than requiring col config
 
-def fix_datetime_columns(
-    df: pd.DataFrame, column_config: Optional[dict]
-) -> pd.DataFrame:
+
+def fix_datetime_columns(df: pd.DataFrame) -> pd.DataFrame:
     """
     Fix datetime columns in a DataFrame based on column configuration.
     """
-    if not column_config:
-        return df
+    df = df.copy(deep=True)
 
-    for column_name, config in column_config.items():
-        col_type = config["type_config"]["type"]
-        if col_type == "datetime":
-            # Convert milliseconds from epoch to datetime
-            df[column_name] = pd.to_datetime(df[column_name], unit="ms")
-        elif col_type == "date":
-            # Convert milliseconds from epoch to date
-            df[column_name] = pd.to_datetime(df[column_name], unit="ms").dt.date
-        elif col_type == "time":
-            # For time values that are already strings in HH:MM:SS format
-            if df[column_name].dtype == "object":
-                df[column_name] = pd.to_datetime(
-                    df[column_name], format="%H:%M:%S"
-                ).dt.time
-            else:
-                # For time values stored as milliseconds since midnight
-                df[column_name] = pd.to_datetime(df[column_name], unit="ms").dt.time
+    for col in df.columns:
+        for idx in df.index:
+            value = df.at[idx, col]
+            if isinstance(value, str):
+                if value.startswith(DATAEDITOR_DATE_VALUE_PREFIX):
+                    df.at[idx, col] = pd.to_datetime(
+                        value.replace(DATAEDITOR_DATE_VALUE_PREFIX, "")
+                    ).date()
+                elif value.startswith(DATAEDITOR_DATETIME_VALUE_PREFIX):
+                    df.at[idx, col] = pd.to_datetime(
+                        value.replace(DATAEDITOR_DATETIME_VALUE_PREFIX, "")
+                    )
+                elif value.startswith(DATAEDITOR_TIME_VALUE_PREFIX):
+                    df.at[idx, col] = pd.to_datetime(
+                        value.replace(DATAEDITOR_TIME_VALUE_PREFIX, ""), format="%H:%M"
+                    ).time()
 
     return df
 
@@ -47,7 +48,7 @@ class DataEditorHandler(WidgetHandler):
 
         # Add column_config to to session state, sinec it is not part of the data
         st.session_state[
-            f"STREAMLIT_PERMALINK_DATA_EDITOR_COLUMN_CONFIG_{self.url_key}"
+            f"{DATAEDITOR_COLUMN_CONFIG_PREFIX}{self.url_key}"
         ] = self.bound_args.arguments.get("column_config")
 
     # Override the url_init method to set the initial fromt he data rather than return
@@ -55,7 +56,7 @@ class DataEditorHandler(WidgetHandler):
         """
         Initialize the URL value(s) in the query params.
         """
-        st.session_state[f"STREAMLIT_PERMALINK_DATA_EDITOR_{self.url_key}"] = (
+        st.session_state[f"{DATAEDITOR_PREFIX}{self.url_key}"] = (
             self.bound_args.arguments.get("data")
         )
         if self.init_url:
@@ -66,6 +67,20 @@ class DataEditorHandler(WidgetHandler):
         # Process URL value: ensure single value and convert to boolean
         parsed_value = self.validate_single_url_value(self.url_value, allow_none=False)
         df = pd.read_json(StringIO(parsed_value), orient="records")
-        df = fix_datetime_columns(df, self.bound_args.arguments.get("column_config"))
-        st.session_state[f"STREAMLIT_PERMALINK_DATA_EDITOR_{self.url_key}"] = df
+        df = fix_datetime_columns(df)
+        st.session_state[f"{DATAEDITOR_PREFIX}{self.url_key}"] = df
         self.bound_args.arguments["data"] = df
+
+    @classmethod
+    def verify_update_url_value(cls, value: Any) -> Any:
+        if not isinstance(value, pd.DataFrame):
+            raise ValueError(
+                f"Data value must be a pandas DataFrame, got {type(value)}"
+            )
+        return value
+
+    @classmethod
+    def verify_get_url_value(cls, value: Any) -> Any:
+        parsed_value = validate_single_url_value(value, allow_none=False)
+        df = pd.read_json(StringIO(parsed_value), orient="records")
+        return [df]